requirements.txt
pyautogui
pynput
dxcam
numpy
Pillow